fixed SignShareRequestHandler for conductor
